Add support for $GNUNET_BASE_CONFIG
The environment variable GNUNET_BASE_CONFIG allows
specifying the directory from which all base config
files will be loaded.  This is useful when ARM is
to be used to supervise things otherwise unrelated to
GNUnet.  Modifying GNUNET_PREFIX is not sufficient
for this use case, since e.g. libexec files need
to be under GNUNET_PREFIX, and GNUNET_BASE_CONFIG
provides a way to only modify the base config
directory independent of where GNUnet was installed.


git-svn-id: https://gnunet.org/svn/gnunet@36919 140774ce-b5e7-0310-ab8b-a85725594a96