Add Vagrant file for development
This patch adds a vagranfile that installs the required
dependencies to compile gnunet.

To be able to use it you need to cd to the contrib/vagrant
directory, run "vagrant up" and wait until the VM is provisioned.
Once it finish the provisioning you can login in the VM with
 "vagrant ssh"  The VM mounts the source code of gnunet in the
folder /gnunet, thus you can edit your code in the host and
compile it in the VM.

Signed-off-by: Antonio Ojea <aojea@midokura.com>